requirements.txt
# LangGraph Writing Assistant Dependencies

# Core LangGraph/LangChain
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0

# Retry logic
tenacity>=8.0.0

# Environment management
python-dotenv>=1.0.0

# Type validation
pydantic>=2.0.0
